working on pathing issue
